update project to support two featueres: did and vc-zkp
- [CORE-1524]
