+++ src/utils/mod.rs
@@ -14,8 +14,11 @@
   limitations under the License.
 */
 
+#[cfg(feature = "did")]
 pub mod substrate;
+#[cfg(feature = "vc-zkp")]
 pub mod utils;
 
+#[cfg(feature = "did")]
 #[macro_use]
 pub mod extrinsic;
